WIP Новый парсер

+++ Threats.Data/Entities/Negative/NegativeType.cs
@@ -2,7 +2,7 @@ namespace Threats.Data.Entities;
 
 public class NegativeType : Entity
 {
-    public NegativeType(int id, string name) : base(id)
+    public NegativeType(long id, string name) : base((int)id)
     {
         Name = name;
     }
